+++ src/plone/jsonapi/routes/datamanagers.py
@@ -130,6 +130,48 @@ class DexterityDataManager(object):
         """
         return IObject.providedBy(field)
 
+    def get_filename(self, **kw):
+        """ extract the filename from the keywords
+        """
+        if "filename" not in kw:
+            logger.debug("ATDataManager::get_filename:No Filename detected"
+                         "-- using title or id")
+            kw["filename"] = kw.get("id") or kw.get("title")
+        return kw.get("filename")
+
+    def get_content_type(self, **kw):
+        """ extract the mimetype from the keywords
+        """
+        if "mimetype" in kw:
+            return kw.get("mimetype")
+        if "content_type" in kw:
+            # same key as in JSON response
+            return kw.get("content_type")
+        return None
+
+    def can_write(self, field):
+        """ check if the field is writeable
+        """
+        # check if the field is read only
+        if field.readonly:
+            raise Unauthorized("Field '%s' is read-only" % field.__name__)
+
+        # XXX: How to check security on the field level?
+        sm = getSecurityManager()
+        permission = permissions.ModifyPortalContent
+        if not sm.checkPermission(permission, self.context):
+            raise Unauthorized("Not allowed to modify this content")
+        return field
+
+    def can_read(self, field):
+        """ check if the field is readable
+        """
+        # XXX: How to check security on the field level?
+        sm = getSecurityManager()
+        if not sm.checkPermission(permissions.View, self.context):
+            raise Unauthorized("Not allowed to view this content")
+        return field
+
     def set(self, name, value, **kw):
         """ Set the field to the given value.
 
@@ -142,28 +184,33 @@ class DexterityDataManager(object):
         if not field:
             return False
 
-        # check if the field is read only
-        if field.readonly:
-            raise Unauthorized("Field '%s' is read-only" % name)
-
-        # XXX: How to check security on the field level?
-        sm = getSecurityManager()
-        permission = permissions.ModifyPortalContent
-        if not sm.checkPermission(permission, self.context):
-            raise Unauthorized("Not allowed to modify this content")
+        # Check the write permission of the field
+        self.can_write(field)
 
         if self.is_file_field(field):
             logger.debug("DexterityDataManager::set:File field"
                          "detected ('%r'), base64 decoding value", field)
             data = str(value).decode("base64")
-            if "filename" not in kw:
-                logger.debug("ATDataManager::set:No Filename detected"
-                             "-- using title or id")
-                kw["filename"] = kw.get("id") or kw.get("title")
-            value = field._type(data=data, filename=kw.get("filename"))
+            filename = self.get_filename(**kw)
+            contentType = self.get_content_type(**kw)
+
+            if contentType:
+                # create NamedFile with content type information
+                value = field._type(data=data, contentType=contentType,
+                                    filename=filename)
+            else:
+                # create NamedFile w/o content type information
+                # -> will be guessed by the extension of the filename
+                value = field._type(data=data, filename=filename)
+
+        # only possible with non file fields
+        if not self.is_file_field(field):
+            # validate the field value
+            field.validate(value)
 
         logger.debug("DexterityDataManager::set:"
                      "field=%r, value=%r", field, value)
+
         field.set(self.context, value)
         return True
 
@@ -171,8 +218,6 @@ class DexterityDataManager(object):
         """ get the value of the field by name
         """
         field = self.get_field(name)
-        # XXX: How to check security on the field level?
-        sm = getSecurityManager()
-        if not sm.checkPermission(permissions.View, self.context):
-            raise Unauthorized("Not allowed to view this content")
+        # Check the read permission of the field
+        self.can_read(field)
         return field.get(self.context)